requirements.txt
amqp==5.0.6
APScheduler==3.6.3
billiard==3.6.4.0
black==21.10b0
cachetools==4.2.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
emoji-country-flag
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
kombu==5.2.0
llvmlite==0.37.0
matplotlib==3.4.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
numba==0.54.1
numpy==1.20.3
pandas==1.3.4
pathspec==0.9.0
patsy==0.5.2
Pillow==8.4.0
pip==21.3.1
platformdirs==2.4.0
prompt-toolkit==3.0.22
protobuf==3.19.1
pyod==0.9.5
pyparsing==3.0.5
python-dateutil==2.8.2
python-dotenv==0.19.1
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
regex==2021.11.2
requests==2.26.0
scikit-learn==1.0.1
scipy==1.6.1
setuptools==58.2.0
six==1.16.0
statsmodels==0.13.0
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.0
